+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `--available`: List the available versions.
   - `--limit`: The maximum number of available versions to list in reverse chronological order, with default value `10`.
 
+- `dfxvm` now supports installing dfx with the aarch64-Darwin binaries.
+
 ## [1.0.0] - 2024-02-20
 
 ## [0.3.1] - 2024-02-07